rocm33-tidy-fixes(09) Fix "narrowing conversion from 'unsigned long' to 'float'".

+++ src/include/miopen/generic_search.hpp
@@ -209,10 +209,10 @@ class HeartBeat
         if(elapsed > 3000)
         {
             elapsed_cumulative += elapsed;
-            const float eta_sec = n_recent != 0u
-                                      ? static_cast<float>((n_total - n_recent) *
-                                                           (elapsed_cumulative / n_recent) / 1000)
-                                      : 0.0f; // paraniod
+            const float eta_sec =
+                n_recent != 0u ? (static_cast<float>(n_total - n_recent) *
+                                  (elapsed_cumulative / static_cast<float>(n_recent)) / 1000.0f)
+                               : 0.0f; // paraniod
             MIOPEN_LOG_W(n_recent << '/' << n_failed << '/' << n_total << ' ' << total_best
                                   << ", best within recent "
                                   << n_within_beat
